refactor: unify readline and signal handling in line_reader module
- Combine readline input and signal handling into line_reader.c
- Add setup_line_reader() for centralized initialization of signals and readline
- Remove separate shell_init and signal_handler modules from main.c
- Move arg_check() and ensure_interactive_mode() as static funcs inside main.c
- Simplify build system: rename read_user_input.c to line_reader.c, remove shell_init.h and signal_handler.h
- Consolidate related logic for improved cohesion and maintainability
