--- inc/readline/read_user_input.h
@@ -1,18 +1,0 @@
-/* ************************************************************************** */
-/*                                                                            */
-/*                                                        :::      ::::::::   */
-/*   read_user_input.h                                  :+:      :+:    :+:   */
-/*                                                    +:+ +:+         +:+     */
-/*   By: didguill <didguill@student.42.fr>          +#+  +:+       +#+        */
-/*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2025/08/03 18:28:27 by didguill          #+#    #+#             */
-/*   Updated: 2025/08/06 15:14:00 by didguill         ###   ########.fr       */
-/*                                                                            */
-/* ************************************************************************** */
-
-#ifndef READ_USER_INPUT_H
-# define READ_USER_INPUT_H
-
-char	*read_user_input(void);
-
-#endif
