--- inc/utils/shell_init.h
@@ -1,18 +1,0 @@
-/* ************************************************************************** */
-/*                                                                            */
-/*                                                        :::      ::::::::   */
-/*   shell_init.h                                       :+:      :+:    :+:   */
-/*                                                    +:+ +:+         +:+     */
-/*   By: didguill <didguill@student.42.fr>          +#+  +:+       +#+        */
-/*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2025/08/03 18:42:14 by didguill          #+#    #+#             */
-/*   Updated: 2025/08/06 16:00:30 by didguill         ###   ########.fr       */
-/*                                                                            */
-/* ************************************************************************** */
-
-#ifndef SHELL_INIT_H
-# define SHELL_INIT_H
-
-void		initialize_shell(int argc);
-
-#endif
