--- inc/utils/signal_handler.h
@@ -1,18 +1,0 @@
-/* ************************************************************************** */
-/*                                                                            */
-/*                                                        :::      ::::::::   */
-/*   signal_handler.h                                   :+:      :+:    :+:   */
-/*                                                    +:+ +:+         +:+     */
-/*   By: didguill <didguill@student.42.fr>          +#+  +:+       +#+        */
-/*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2025/08/06 16:06:51 by didguill          #+#    #+#             */
-/*   Updated: 2025/08/06 16:07:14 by didguill         ###   ########.fr       */
-/*                                                                            */
-/* ************************************************************************** */
-
-#ifndef SIGNAL_HANDLER_H
-# define SIGNAL_HANDLER_H
-
-void	setup_signal_handlers(void);
-
-#endif
